feat: collection_settingsのCRUD完成

+++ src/models/collection_setting.cr
@@ -9,7 +9,7 @@ class CollectionSetting < BaseModel
     # 機関コード(コレクション名の略号)
     column institution_code : String = "", allow_blank: true
     # 標本IDの最終番号
-    column latest_collection_code : Int64 = 0i64
+    column latest_collection_code : Int32 = 0
     # 備考
     column note : String = "", allow_blank: true
   end
